requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==4.2.18
django-cors-headers==4.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
joblib==1.4.2
Levenshtein==0.27.1
nltk==3.9.1
PyJWT==2.9.0
python-Levenshtein==0.27.1
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
